requirements.txt
discord.py==2.3.2
ossapi==3.4.3
python-dotenv==1.0.1
mcrcon==0.7.0